Allows external SVGs (#953)
* Make Icon component stateless

* Add iconsPath Global setting

* Add path prop to icon-related components

to allow per icon folder paths to provided SLDS icons

* Add storybook middleware

to allow icons to be at /assets/icons in order to match SLDS markup

* Add icon examples and snapshots

* Add classNames wrapper

to prevent output of class=“”

* External URL Icon markup update

- Allow inline icon objects to to still be passed in
- Move icon stories to own files
- Simplify classnames and update to 2.2.1

* Update tests related to external URL Icon changes

Make snapshot tests closer to SLDS markup

* Update snapshots for external icons

and classname being undefined instead of “”

* Fix GlobalHeader story

* Prevent assistiveText from being passed onto SVG

* Fix eslint issues in button icon and button

* Add checkProps to detect if icon is missing

And warn how to fix if URL doesn’t load

* Update Icon tests to use enzyme

* Update icon tests page header/notifications

* Remove icon classname string template usage

To allow better find and replace for class names

* Icon: Make path override global settings and be full URL

* Icon: Make path override global settings and be full URL

* Add more conditions for url exists checkProp

Using XMLHttpRequest can cause problems in non-browser environments. This should be completely removed in production environment and should not execute in a testing environment.